// src-31052018-currently-working/app/plan/enums/plan-action.ts
/**
 * Enumeration to represent actions to perform on a plan
 */
export enum PlanAction {
  COPY = 'Copy Plan',
  CANCEL = 'Cancel Plan',
  DELETE = 'Delete Plan',
  EXPORT_DATA_DETAIL = 'Export Data Detail',
  REINSTATE = 'Reinstate Plan',
  RETRO_REVISE = 'Retro-Revise Plan',
  REVERT = 'Revert Plan',
  REVISE = 'Revise Plan',
  TERMINATE = 'Terminate Plan',
  EXPORT_COMMENTS = 'Export Comments'
}